Fixing reading of incomplete projections
Analysis:
If the raw data recording failed partway through (for example,
if RTHawk crashed during data recording), the size of the last
projection may not be correct. Currently, readRthRaw will
exit unexpectedly instead of reconstructing the correct projections.

Resolution:
Checking the size of the read projection and breaking out of
read loop if it is less than the expected size. The previous
projections are processed and plotted.