Adding option to save all plots to files
Analysis:
The only way to save plots to image files currently is to
use the "Save" button for each individual plot. For batch
conversion of plot frames to video, this is not convenient.

Resolution:
Adding an option to batch save all the projection plots to
individual png files. Any argument after the first one will
run the batch saving and not show the GUI.

Note that the output filenames are hard-coded to proj%04d.png.
This could be improved.