Adding saving of peaks to coordinate text files
The "-c" / "--coord-save" option can now be used to save the
peak coordinate positions to a text file. The plot png and coordinate
output can be combined but neither of these options can be combined
with the GUI option.

Note that the coordinate text file is saved to the same directory
as the original projection file, while the plots are saved to
the current directory.