Adding snr calculation
Analysis:
Want a way to output SNR values for projections.

Resolution:
Calculating SNR using the same algorithm as is used in the peakfinder recon.
This has been implemented in a separate snrCalc module.

SNR's are calculated in magnitude mode and the minimum SNR of the three
projections is output in saveCoords mode right after the coordinates.