HAWK-29 Reading and writing timestamps from latest raw files
Analysis:
Now that the CathExport plugin writes timestamps into the raw
projection file, these must be read by readRthRaw.

Resolution:
Reading timestamp data from the raw data file. These are also
written to the coordinate file if this option is selected.

The timestamp is written after the coordinates and SNR and
before the physiological data. The timestamp represents
the number of milliseconds since Unix epoch.

Older format files can still be read using the "legacy" option
for files that do not contain physiological data or timestamps,
and the "legacy2" option for files that contain physiological data
but no timestamps.