Save peek animation to output file, fixes #9
Added a new "-f" flag to peek to save animation to file
- the frame rate of the animation is set
to 10 fps (for easier inspection)
- the dpi of the animation files is set to 200
for better resolution

The new flag takes an output filename (including extension)
for the animation. Saving animations to files requires
the installation of ffmpeg on the host.

It should work for the following video extensions:
mov, mp4, avi