Adding legends to the peek plots, #10
To help understand which axis a plot represents,
the peek plots for a projection file are now
labeled X Y Z for SRI sequences, and just X
for hadamard sequences (which only save a
single axis to a projeciton file).

Coil numbers are now added to the plots
when using cathy peek for a directory.