Extending hard-coded color map fix to ground truth lines #5
A coil index dictionary for only 4 indices was still being used
for the ground truth line plots. This has now been replaced with
the qualitative colormap.

A comment explains the limitation that coils beyond the range
of the 10-value colormap could result in repeated colors being
used in the plot.