Adding max iterations option to run_localize #23
The default maximum iterations (used for PNG & JPNG)
is set to 32: note that this will change the behaviour
of the default call as JPNG's default max iterations
in the localization module is 256. So the coordinate
output may now be different in JPNG.

PNG always defaulted to 32, so no change should be
seen with this algorithm.

This makes the calls to the PNG & JPNG algorithms more
consistent as they will be called using the same number
of maximum iterations, for a fairer comparison.